Migrate Meilisearch to IndexerInterface and SearcherInterface
