+++ packages/seal-meilisearch-adapter/MeilisearchAdapter.php
@@ -4,22 +4,28 @@ namespace Schranz\Search\SEAL\Adapter\Meilisearch;
 
 use Meilisearch\Client;
 use Schranz\Search\SEAL\Adapter\AdapterInterface;
-use Schranz\Search\SEAL\Adapter\ConnectionInterface;
+use Schranz\Search\SEAL\Adapter\IndexerInterface;
 use Schranz\Search\SEAL\Adapter\SchemaManagerInterface;
+use Schranz\Search\SEAL\Adapter\SearcherInterface;
 
 final class MeilisearchAdapter implements AdapterInterface
 {
-    private readonly ConnectionInterface $connection;
 
     private readonly SchemaManagerInterface $schemaManager;
 
+    private readonly IndexerInterface $indexer;
+
+    private readonly SearcherInterface $searcher;
+
     public function __construct(
         private readonly Client $client,
-        ?ConnectionInterface $connection = null,
         ?SchemaManagerInterface $schemaManager = null,
+        ?IndexerInterface $indexer = null,
+        ?SearcherInterface $searcher = null,
     ) {
-        $this->connection = $connection ?? new MeilisearchConnection($client);
         $this->schemaManager = $schemaManager ?? new MeilisearchSchemaManager($client);
+        $this->indexer = $indexer ?? new MeilisearchIndexer($client);
+        $this->searcher = $searcher ?? new MeilisearchSearcher($client);
     }
 
     public function getSchemaManager(): SchemaManagerInterface
@@ -27,8 +33,13 @@ final class MeilisearchAdapter implements AdapterInterface
         return $this->schemaManager;
     }
 
-    public function getConnection(): ConnectionInterface
+    public function getIndexer(): IndexerInterface
     {
-        return $this->connection;
+        return $this->indexer;
+    }
+
+    public function getSearcher(): SearcherInterface
+    {
+        return $this->searcher;
     }
 }
