+++ packages/seal-meilisearch-adapter/MeilisearchSearcher.php
@@ -4,16 +4,14 @@ namespace Schranz\Search\SEAL\Adapter\Meilisearch;
 
 use Meilisearch\Client;
 use Meilisearch\Exceptions\ApiException;
-use Schranz\Search\SEAL\Adapter\ConnectionInterface;
+use Schranz\Search\SEAL\Adapter\SearcherInterface;
 use Schranz\Search\SEAL\Marshaller\Marshaller;
 use Schranz\Search\SEAL\Schema\Index;
 use Schranz\Search\SEAL\Search\Condition;
 use Schranz\Search\SEAL\Search\Result;
 use Schranz\Search\SEAL\Search\Search;
-use Schranz\Search\SEAL\Task\AsyncTask;
-use Schranz\Search\SEAL\Task\TaskInterface;
 
-final class MeilisearchConnection implements ConnectionInterface
+final class MeilisearchSearcher implements SearcherInterface
 {
     private Marshaller $marshaller;
 
@@ -22,48 +20,6 @@ final class MeilisearchConnection implements ConnectionInterface
     ) {
         $this->marshaller = new Marshaller();
     }
-
-    public function save(Index $index, array $document, array $options = []): ?TaskInterface
-    {
-        $identifierField = $index->getIdentifierField();
-
-        /** @var string|null $identifier */
-        $identifier = ((string) $document[$identifierField->name]) ?? null;
-
-        $indexResponse = $this->client->index($index->name)->addDocuments([
-            $this->marshaller->marshall($index->fields, $document),
-        ], $identifierField->name);
-
-        if ($indexResponse['status'] !== 'enqueued') {
-            throw new \RuntimeException('Unexpected error while save document with identifier "' . $identifier . '" into Index "' . $index->name . '".');
-        }
-
-        if (true !== ($options['return_slow_promise_result'] ?? false)) {
-            return null;
-        }
-
-        return new AsyncTask(function() use ($indexResponse) {
-            $this->client->waitForTask($indexResponse['taskUid']);
-        });
-    }
-
-    public function delete(Index $index, string $identifier, array $options = []): ?TaskInterface
-    {
-        $deleteResponse = $this->client->index($index->name)->deleteDocument($identifier);
-
-        if ($deleteResponse['status'] !== 'enqueued') {
-            throw new \RuntimeException('Unexpected error while delete document with identifier "' . $identifier . '" from Index "' . $index->name . '".');
-        }
-
-        if (true !== ($options['return_slow_promise_result'] ?? false)) {
-            return null;
-        }
-
-        return new AsyncTask(function() use ($deleteResponse) {
-            $this->client->waitForTask($deleteResponse['taskUid']);
-        });
-    }
-
     public function search(Search $search): Result
     {
         // optimized single document query
