--- packages/seal-meilisearch-adapter/Tests/MeilisearchConnectionTest.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace Schranz\Search\SEAL\Adapter\Meilisearch\Tests;
-
-use Schranz\Search\SEAL\Adapter\Meilisearch\MeilisearchConnection;
-use Schranz\Search\SEAL\Adapter\Meilisearch\MeilisearchSchemaManager;
-use Schranz\Search\SEAL\Testing\AbstractConnectionTestCase;
-
-class MeilisearchConnectionTest extends AbstractConnectionTestCase
-{
-    private static $client;
-
-    public static function setUpBeforeClass(): void
-    {
-        self::$client = ClientHelper::getClient();
-
-        self::$connection = new MeilisearchConnection(self::$client);
-        self::$schemaManager = new MeilisearchSchemaManager(self::$client);
-
-        parent::setUpBeforeClass();
-    }
-
-    public function testFindMultipleIndexes(): void
-    {
-        $this->markTestSkipped('Not supported by Meilisearch: https://github.com/schranz-search/schranz-search/issues/28');
-    }
-}
