+++ packages/seal-meilisearch-adapter/composer.json
@@ -32,7 +32,7 @@
     "require": {
         "php": "^8.1",
         "schranz-search/seal": "^0.1",
-        "meilisearch/meilisearch-php": "^1.0.0",
+        "meilisearch/meilisearch-php": "^1.0",
         "psr/container": "^1.0 || ^2.0"
     },
     "require-dev": {
